Adjusts pointRobot according to new structure. Adds sensing to abstractRobot
